Refactor: move FilterHelpers under generic util folder
