+++ src/lib/Components/ArtworkFilterOptions/PriceRangeOptions.tsx
@@ -1,9 +1,9 @@
 import {
-  AggregateOption,
-  FilterDisplayName,
-  FilterParamName,
-} from "lib/Scenes/Collection/Helpers/FilterArtworksHelpers"
-import { ArtworkFilterContext, FilterData, useSelectedOptionsDisplay } from "lib/utils/ArtworkFiltersStore"
+  ArtworkFilterContext,
+  FilterData,
+  useSelectedOptionsDisplay,
+} from "lib/utils/ArtworkFilter/ArtworkFiltersStore"
+import { AggregateOption, FilterDisplayName, FilterParamName } from "lib/utils/ArtworkFilter/FilterArtworksHelpers"
 import React, { useContext } from "react"
 import { NavigatorIOS } from "react-native"
 import { aggregationForFilter } from "../FilterModal"
