+++ src/lib/Components/ArtworkFilterOptions/__tests__/PriceRangeOptions-tests.tsx
@@ -1,12 +1,16 @@
-import { FilterParamName, InitialState } from "lib/Scenes/Collection/Helpers/FilterArtworksHelpers"
 import { extractText } from "lib/tests/extractText"
 import { renderWithWrappers } from "lib/tests/renderWithWrappers"
+import { FilterParamName, InitialState } from "lib/utils/ArtworkFilter/FilterArtworksHelpers"
 import { Box, Theme } from "palette"
 import React from "react"
 import { ReactTestRenderer } from "react-test-renderer"
 import { FakeNavigator as MockNavigator } from "../../../../lib/Components/Bidding/__tests__/Helpers/FakeNavigator"
 import { OptionListItem } from "../../../../lib/Components/FilterModal"
-import { Aggregations, ArtworkFilterContext, ArtworkFilterContextState } from "../../../utils/ArtworkFiltersStore"
+import {
+  Aggregations,
+  ArtworkFilterContext,
+  ArtworkFilterContextState,
+} from "../../../utils/ArtworkFilter/ArtworkFiltersStore"
 import { PriceRangeOptionsScreen } from "../PriceRangeOptions"
 import { InnerOptionListItem } from "../SingleSelectOption"
