+++ src/lib/Scenes/Fair2/Fair2.tsx
@@ -2,7 +2,7 @@ import { Fair2_fair } from "__generated__/Fair2_fair.graphql"
 import { Fair2Query } from "__generated__/Fair2Query.graphql"
 import { AnimatedArtworkFilterButton, FilterModalMode, FilterModalNavigator } from "lib/Components/FilterModal"
 import { defaultEnvironment } from "lib/relay/createEnvironment"
-import { ArtworkFilterContext, ArtworkFilterGlobalStateProvider } from "lib/utils/ArtworkFiltersStore"
+import { ArtworkFilterContext, ArtworkFilterGlobalStateProvider } from "lib/utils/ArtworkFilter/ArtworkFiltersStore"
 import renderWithLoadProgress from "lib/utils/renderWithLoadProgress"
 import { Schema } from "lib/utils/track"
 import { Box, Separator, Spacer, Theme } from "palette"
